Update Group SPD endpoint tests
