+++ explainability-service/src/test/java/org/kie/trustyai/service/endpoints/metrics/fairness/group/GroupStatisticalParityDifferenceEndpointMemoryTest.java
@@ -1,0 +1,68 @@
+package org.kie.trustyai.service.endpoints.metrics.fairness.group;
+
+import java.time.LocalDateTime;
+import java.util.Map;
+
+import org.jboss.resteasy.reactive.RestResponse;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.kie.trustyai.explainability.model.Dataframe;
+import org.kie.trustyai.explainability.model.Prediction;
+import org.kie.trustyai.explainability.model.PredictionMetadata;
+import org.kie.trustyai.explainability.model.SimplePrediction;
+import org.kie.trustyai.service.endpoints.metrics.MetricsEndpointTestProfile;
+import org.kie.trustyai.service.endpoints.metrics.RequestPayloadGenerator;
+import org.kie.trustyai.service.mocks.MockDatasource;
+import org.kie.trustyai.service.mocks.MockMemoryStorage;
+import org.kie.trustyai.service.mocks.MockPrometheusScheduler;
+import org.kie.trustyai.service.payloads.BaseScheduledResponse;
+import org.kie.trustyai.service.payloads.metrics.BaseMetricResponse;
+import org.kie.trustyai.service.payloads.metrics.fairness.group.GroupMetricRequest;
+import org.kie.trustyai.service.payloads.scheduler.ScheduleId;
+import org.kie.trustyai.service.payloads.scheduler.ScheduleList;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+import io.quarkus.test.common.http.TestHTTPEndpoint;
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+import io.restassured.http.ContentType;
+
+import jakarta.enterprise.inject.Instance;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.core.Response;
+import org.kie.trustyai.service.profiles.MemoryTestProfile;
+
+import static io.restassured.RestAssured.given;
+import static io.restassured.RestAssured.when;
+import static org.hamcrest.Matchers.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+@QuarkusTest
+@TestProfile(MemoryTestProfile.class)
+@TestHTTPEndpoint(GroupStatisticalParityDifferenceEndpoint.class)
+class GroupStatisticalParityDifferenceEndpointMemoryTest extends GroupStatisticalParityDifferenceEndpointBaseTest {
+
+    @Inject
+    Instance<MockMemoryStorage> storage;
+
+    @Inject
+    Instance<MockPrometheusScheduler> scheduler;
+
+    @BeforeEach
+    void populateStorage() throws JsonProcessingException {
+        storage.get().emptyStorage();
+        final Dataframe dataframe = datasource.get().generateRandomDataframe(1000);
+        datasource.get().saveDataframe(dataframe, MODEL_ID);
+        datasource.get().saveMetadata(datasource.get().createMetadata(dataframe), MODEL_ID);
+    }
+
+    @AfterEach
+    void clearRequests() {
+        // prevent a failing test from failing other tests erroneously
+        scheduler.get().getAllRequests().clear();
+    }
+
+}
